fix: sankey link path style and axis grid style callback
