+++ packages/vchart/src/component/axis/cartesian/axis.ts
@@ -25,10 +25,9 @@ import type { IComponentOption } from '../../interface';
 // eslint-disable-next-line no-duplicate-imports
 import { ComponentTypeEnum } from '../../interface';
 import { HOOK_EVENT } from '@visactor/vgrammar';
-
 import type { LineAxisAttributes } from '@visactor/vrender-components';
 // eslint-disable-next-line no-duplicate-imports
-import { getAxisAttributes, isValidCartesianAxis } from '../utils';
+import { isValidCartesianAxis } from '../utils';
 import type { IAxis, ITick } from '../interface';
 import { registerDataSetInstanceParser, registerDataSetInstanceTransform } from '../../../data/register';
 import { scaleParser } from '../../../data/parser/scale';
@@ -249,7 +248,7 @@ export abstract class CartesianAxis extends AxisComponent implements IAxis {
   setAttrFromSpec() {
     super.setAttrFromSpec();
 
-    const axisStyle: any = getAxisAttributes(this._spec, this._theme);
+    const axisStyle: any = this._getAxisAttributes();
     axisStyle.label.formatMethod = this.getLabelFormatMethod();
     axisStyle.verticalFactor = this.orient === 'top' || this.orient === 'right' ? -1 : 1;
     this._axisStyle = axisStyle;
