+++ packages/vchart/src/theme/buildin-theme/light/series/sankey.ts
@@ -1,0 +1,12 @@
+import type { ISankeySeriesTheme } from '../../../../series/sankey/interface';
+
+export const sankey: ISankeySeriesTheme = {
+  link: {
+    style: {
+      lineWidth: 0,
+      fillOpacity: 0.15,
+      direction: 'horizontal',
+      round: true
+    }
+  }
+};
